+++ packages/apollo-link-http/src/httpLink.ts
@@ -1,6 +1,6 @@
 import { ApolloLink, Observable, RequestHandler, fromError } from 'apollo-link';
 import {
-  serializeFetchBody,
+  serializeFetchParameter,
   selectURI,
   parseAndCheckHttpResponse,
   checkFetcher,
@@ -48,7 +48,7 @@ export const createHttpLink = (linkOptions: HttpLink.Options = {}) => {
   };
 
   return new ApolloLink(operation => {
-    const chosenURI = selectURI(operation, uri);
+    let chosenURI = selectURI(operation, uri);
 
     const context = operation.getContext();
 
@@ -70,10 +70,67 @@ export const createHttpLink = (linkOptions: HttpLink.Options = {}) => {
     const { controller, signal } = createSignalIfSupported();
     if (controller) (options as any).signal = signal;
 
-    try {
-      (options as any).body = serializeFetchBody(body);
-    } catch (parseError) {
-      return fromError(parseError);
+    if (options.method === 'GET') {
+      // Implement the standard HTTP GET serialization, plus 'extensions'. Note
+      // the extra level of JSON serialization!
+      const queryParams = [];
+      const addQueryParam = (key: string, value: string) => {
+        queryParams.push(`${key}=${encodeURIComponent(value)}`);
+      };
+
+      if ('query' in body) {
+        addQueryParam('query', body.query);
+      }
+      if (body.operationName) {
+        addQueryParam('operationName', body.operationName);
+      }
+      if (body.variables) {
+        let serializedVariables;
+        try {
+          serializedVariables = serializeFetchParameter(
+            body.variables,
+            'Variables map',
+          );
+        } catch (parseError) {
+          return fromError(parseError);
+        }
+        addQueryParam('variables', serializedVariables);
+      }
+      if (body.extensions) {
+        let serializedExtensions;
+        try {
+          serializedExtensions = serializeFetchParameter(
+            body.extensions,
+            'Extensions map',
+          );
+        } catch (parseError) {
+          return fromError(parseError);
+        }
+        addQueryParam('extensions', serializedExtensions);
+      }
+
+      // Reconstruct the URI with added query params.
+      // XXX This assumes that the URI is well-formed and that it doesn't
+      //     already contain any of these query params. We could instead use the
+      //     URL API and take a polyfill (whatwg-url@6) for older browsers that
+      //     don't support URLSearchParams. Note that some browsers (and
+      //     versions of whatwg-url) support URL but not URLSearchParams!
+      let fragment = '',
+        preFragment = chosenURI;
+      const fragmentStart = chosenURI.indexOf('#');
+      if (fragmentStart !== -1) {
+        fragment = chosenURI.substr(fragmentStart);
+        preFragment = chosenURI.substr(0, fragmentStart);
+      }
+      const queryParamsPrefix = preFragment.indexOf('?') === -1 ? '?' : '&';
+      chosenURI =
+        preFragment + queryParamsPrefix + queryParams.join('&') + fragment;
+    } else {
+      try {
+        (options as any).body = serializeFetchParameter(body, 'Payload');
+      } catch (parseError) {
+        return fromError(parseError);
+      }
     }
 
     return new Observable(observer => {