+++ packages/apollo-link-http/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change log
 
 ### vNEXT
-- move logic to apollo-link-utilities [PR#364](https://github.com/apollographql/apollo-link/pull/364)
+- move logic to apollo-link-http-core [PR#364](https://github.com/apollographql/apollo-link/pull/364)
 
 ### 1.3.3
 - ApolloLink upgrade